fix: change execution_order column type from Integer to Float
ROOT CAUSE OF REORDERING ISSUE FOUND:
- execution_order column was defined as Column(Integer)
- SQLAlchemy was truncating decimal values (0.5 → 0, 1.5 → 1)
- Database never received the fractional order values
- Scripts stayed in same position after reorder

Fix:
- Change Column(Integer) to Column(Float) in RepositoryScript model
- Now preserves decimal values like 0.5, 1.5, 2.5
- SQLite is dynamically typed so no migration needed
- Existing INTEGER column will accept FLOAT values

How reordering works:
- Scripts at orders: [1, 2, 3]
- Move script 2 up → newOrder = 2 - 1.5 = 0.5
- Save 0.5 to database (previously truncated to 0)
- Fetch and sort: [0.5, 1, 3]
- Script with order 0.5 appears first ✓

After restart, reordering should work correctly!

+++ app/database/models.py
@@ -373,7 +373,7 @@ class RepositoryScript(Base):
 
     # Hook configuration
     hook_type = Column(String(50), nullable=False)  # 'pre-backup', 'post-backup'
-    execution_order = Column(Integer, default=1, nullable=False)  # Order in chain (1, 2, 3...)
+    execution_order = Column(Float, default=1, nullable=False)  # Order in chain (supports decimals for reordering)
     enabled = Column(Boolean, default=True, nullable=False)
 
     # Per-repository overrides
